backend/requirements.txt
fastapi==0.123.5
uvicorn==0.30.1
sqlmodel==0.0.19
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg==3.1.18